To fiber

+++ Entities/Message.go
@@ -3,7 +3,7 @@ package Entities
 // ID is a primary key
 type Message struct {
 	//Primary key
-	ID int64 `json:"id"`
+	ID uint `json:"id"`
 	//Index
 	ParentChatID    uint     `json:"parent_chat_id"`
 	ParentUserID    uint     `json:"parent_user_id"`
@@ -14,7 +14,7 @@ type Message struct {
 }
 
 type MessageDTO struct {
-	ID              int64    `json:"id"`
+	ID              uint     `json:"id"`
 	ParentChatID    uint     `json:"parent_chat_id" binding:"required"`
 	ParentUserID    uint     `json:"parent_user_id" binding:"required"`
 	Data            string   `json:"data" binding:"required"`
